pyproject.toml
[tool.poetry]
name = "chrome-local-storage"
description = "Interact with Google Chrome local storage"
version = "0.5.1"
license = "MIT-0"
authors = ["Judson Neer <judson.neer@gmail.com>"]
repository = "https://github.com/lordjabez/chrome-local-storage"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7 <3.10"
requests = "^2.25.0"
trio-chrome-devtools-protocol = "^0.6.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.4"
flake8 = "^4.0.1"
pytest = "^7.1.0"
pytest-cov = "^3.0.0"
safety = "^1.10.3"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
